+++ tests/unit-tests/point/boundary_tests.cpp
@@ -1,6 +1,7 @@
 #include "enumerations/interface.hpp"
 #include "specfem/point/boundary.hpp"
 #include "specfem_setup.hpp"
+#include "test_helper.hpp"
 #include "test_macros.hpp"
 #include "utilities/simd.hpp"
 #include <Kokkos_Core.hpp>
@@ -30,12 +31,6 @@ protected:
       Kokkos::finalize();
   }
 };
-
-// For better naming
-struct Serial : std::integral_constant<bool, false> {};
-struct SIMD : std::integral_constant<bool, true> {};
-
-using TestTypes = ::testing::Types<Serial, SIMD>;
 
 template <typename T>
 class PointBoundaryTest : public PointBoundaryTestUntyped<T::value> {};
