+++ tests/unit-tests/point/jacobian_matrix_tests.cpp
@@ -4,6 +4,7 @@
 #include "enumerations/interface.hpp"
 #include "specfem/point/jacobian_matrix.hpp"
 #include "specfem_setup.hpp"
+#include "test_helper.hpp"
 #include "test_macros.hpp"
 #include "utilities/simd.hpp"
 #include <Kokkos_Core.hpp>
@@ -32,12 +33,6 @@ protected:
       Kokkos::finalize();
   }
 };
-
-// For better naming
-struct Serial : std::integral_constant<bool, false> {};
-struct SIMD : std::integral_constant<bool, true> {};
-
-using TestTypes = ::testing::Types<Serial, SIMD>;
 
 template <typename T>
 class PointJacobianMatrixTest
