+++ tests/unit-tests/point/properties/properties_tests.hpp
@@ -1,5 +1,7 @@
+#pragma once
 #include "datatypes/simd.hpp"
 #include "specfem_setup.hpp"
+#include "test_helper.hpp"
 #include <Kokkos_Core.hpp>
 #include <gtest/gtest.h>
 
@@ -22,12 +24,6 @@ protected:
       Kokkos::finalize();
   }
 };
-
-// For better naming
-struct Serial : std::integral_constant<bool, false> {};
-struct SIMD : std::integral_constant<bool, true> {};
-
-using TestTypes = ::testing::Types<Serial, SIMD>;
 
 template <typename T>
 class PointPropertiesTest : public PointPropertiesTestUntyped<T::value> {};
